+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - By default, add bracket spacing for maps ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/99)).
 - Fix extraneous linebreaks being added after `this` variable expression ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/98)).
+- Remove `--apex-anonymous` option, use `apex-anonymous` parser instead ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/106)).
 
 # 1.0.0-rc.6